refactor!: rename directory of `composables` to `utils` for the sake of common sense
