+++ resources/client/shim.d.ts
@@ -3,7 +3,7 @@ import ziggyRoute from 'ziggy-js'
 import type { PageProps, Page } from '@inertiajs/core'
 import type { RouteParam, RouteParamsWithQueryOverload, Config } from 'ziggy-js'
 import type { AppRoutes } from '~/modules/ziggy'
-import type { NavigationType, NavigationItem } from '~/composables/navigations'
+import type { NavigationType, NavigationItem } from '~/utils/navigations'
 
 export {}
