+++ resources/client/utils/common.ts
